Quiz setup and dummy database
The three quiz types (multiple choices, true or false, and fill in the blanks) are set up here.
Also added a dummy database

Note:
Still need to integrate the 3 quiz types and add more features to have it work completely. Also, still have to update the database to something that really makes sense